Use XdgConfig instead of XdgData
In the spirit of the original FHS (filesystem hierarchy standard), datadir
is for read-only platform-independent data.
Before this patch, mywork creates projects.json in
~/.local/share/mywork/
which looks just wrong.  ~/.local mostly adheres to the FHS, which means the
share subdir is ment to be read-only.
Since this is configuration data, it should go to
~/.config/mywork/